Change license to BSD-3-Clause
Prior to this change, the project was using the Clear BSD License.
However, pkg.go.dev does not recognize that, and it is an issue with
displaying README and other functionality.

This change simply modifies the license to standard BSD-3-Clause, which
is the same as BSD-3-Clause-Clear without the patent license text.