Make device an interface
Prior to this change, the context would store the pointer to the
usb.Device. While the type itself is valid, it prevents us from
validating the vendor specific control packets we need in an automated
fashion.

This change converts the device member to be an interface of type
`usbDevice`. This is an internal interface intended to mimic the libusb
functions used by the x52 library.