Retry device action on failure
Prior to this change, the assumption was that gousb would handle all
retries internally. However, this appears to be true only for a small
subset of the possible error cases, such as transaction interrupted. For
other errors such as pipe error, gousb returns the error directly, and
we need to retry the action ourselves.

This change wraps the call to the gousb methods in a retry loop and
returns on the first success. What I've observed is that even if one
control message fails, it usually succeeds on retry. To be on the safe
side, I've allowed the requests to be retried up to 3 times before
failing.